Convert EditTagsScreen to EditTagsDialog and add delete option to multiselect toolbar
